pyproject.toml
[project]
name = "finances-sync"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "actualpy>=0.5.0",
    "httpx>=0.27.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
